03-Langchian/03-实战项目/02-Rag实战/rag-project/requirements.txt
langchain>=0.0.267
openai>=0.27.8
chromadb>=0.4.6
sentence-transformers>=2.2.2
pypdf>=3.15.1
tiktoken>=0.4.0
fastapi>=0.100.0
uvicorn>=0.23.1
python-dotenv>=1.0.0
pydantic>=2.0.0 
langchain_community==0.3.20